BUG FIX: Troca do tipo float para double e int (apenas as variáveis que manipulam os PIs) para unsigned long long int. Motivo: float e int estouravam pra matrizes muito grandes.

+++ src/sequencial.c
@@ -18,7 +18,7 @@ Resultado* resultadoFinal;
 Resultado* sequencial(int k) {
 
 	resultadoFinal = (Resultado*)malloc(sizeof(Resultado));
-	resultadoFinal->PI = (int*)malloc(k * sizeof(int));
+	resultadoFinal->PI = (unsigned long long int*)malloc(k * sizeof(unsigned long long int));
 
 	Tarefas *tarefas = (Tarefas*)malloc(sizeof(Tarefas));
 	tarefas->inicio = 0;
