+++ src/tarefa.h
@@ -19,7 +19,7 @@ Tarefas* divideTarefas(int undExecucao);
 recebe i, a linha e coluna em cima de onde calcular o PI e referencias para o maior e menor elemento
 retorna o valor do PI e, por referencia os valores do maior e menor elemento
 */
-int produtoInternoEMaiorMenorElementoParcial(int i, int* maiorElem, int* menorElem);
+unsigned long long int produtoInternoEMaiorMenorElementoParcial(int i, int* maiorElem, int* menorElem);
 
 /*
 recebe vetor contendo os indices das linhas/colunas das quais
